# backend/ent/paytype.go
// Code generated by entc, DO NOT EDIT.

package ent

import (
	"fmt"
	"strings"

	"github.com/facebookincubator/ent/dialect/sql"
	"github.com/theuo/app/ent/paytype"
)

// PayType is the model entity for the PayType schema.
type PayType struct {
	config `json:"-"`
	// ID of the ent.
	ID int `json:"id,omitempty"`
	// TypeInfo holds the value of the "TypeInfo" field.
	TypeInfo string `json:"TypeInfo,omitempty"`
	// Edges holds the relations/edges for other nodes in the graph.
	// The values are being populated by the PayTypeQuery when eager-loading is set.
	Edges PayTypeEdges `json:"edges"`
}

// PayTypeEdges holds the relations/edges for other nodes in the graph.
type PayTypeEdges struct {
	// Payments holds the value of the payments edge.
	Payments []*Payment
	// loadedTypes holds the information for reporting if a
	// type was loaded (or requested) in eager-loading or not.
	loadedTypes [1]bool
}

// PaymentsOrErr returns the Payments value or an error if the edge
// was not loaded in eager-loading.
func (e PayTypeEdges) PaymentsOrErr() ([]*Payment, error) {
	if e.loadedTypes[0] {
		return e.Payments, nil
	}
	return nil, &NotLoadedError{edge: "payments"}
}

// scanValues returns the types for scanning values from sql.Rows.
func (*PayType) scanValues() []interface{} {
	return []interface{}{
		&sql.NullInt64{},  // id
		&sql.NullString{}, // TypeInfo
	}
}

// assignValues assigns the values that were returned from sql.Rows (after scanning)
// to the PayType fields.
func (pt *PayType) assignValues(values ...interface{}) error {
	if m, n := len(values), len(paytype.Columns); m < n {
		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
	}
	value, ok := values[0].(*sql.NullInt64)
	if !ok {
		return fmt.Errorf("unexpected type %T for field id", value)
	}
	pt.ID = int(value.Int64)
	values = values[1:]
	if value, ok := values[0].(*sql.NullString); !ok {
		return fmt.Errorf("unexpected type %T for field TypeInfo", values[0])
	} else if value.Valid {
		pt.TypeInfo = value.String
	}
	return nil
}

// QueryPayments queries the payments edge of the PayType.
func (pt *PayType) QueryPayments() *PaymentQuery {
	return (&PayTypeClient{config: pt.config}).QueryPayments(pt)
}

// Update returns a builder for updating this PayType.
// Note that, you need to call PayType.Unwrap() before calling this method, if this PayType
// was returned from a transaction, and the transaction was committed or rolled back.
func (pt *PayType) Update() *PayTypeUpdateOne {
	return (&PayTypeClient{config: pt.config}).UpdateOne(pt)
}

// Unwrap unwraps the entity that was returned from a transaction after it was closed,
// so that all next queries will be executed through the driver which created the transaction.
func (pt *PayType) Unwrap() *PayType {
	tx, ok := pt.config.driver.(*txDriver)
	if !ok {
		panic("ent: PayType is not a transactional entity")
	}
	pt.config.driver = tx.drv
	return pt
}

// String implements the fmt.Stringer.
func (pt *PayType) String() string {
	var builder strings.Builder
	builder.WriteString("PayType(")
	builder.WriteString(fmt.Sprintf("id=%v", pt.ID))
	builder.WriteString(", TypeInfo=")
	builder.WriteString(pt.TypeInfo)
	builder.WriteByte(')')
	return builder.String()
}

// PayTypes is a parsable slice of PayType.
type PayTypes []*PayType

func (pt PayTypes) config(cfg config) {
	for _i := range pt {
		pt[_i].config = cfg
	}
}